+++ plugins/sitepress-multilingual-cms/embedded/autoload.php
@@ -4,4 +4,4 @@
 
 require_once __DIR__ . '/composer' . '/autoload_real.php';
 
-return ComposerAutoloaderIniteb8f8377d47a308a5553b21cf94141d2::getLoader();
+return ComposerAutoloaderInit27c2a16dc375eff11ab5003351ac611b::getLoader();