+++ plugins/wp-fastest-cache/inc/admin.php
@@ -333,11 +333,14 @@
 			// if(defined('DONOTCACHEPAGE')){
 			// 	return array("DONOTCACHEPAGE <label>constant is defined as TRUE. It must be FALSE</label>", "error");
 			// }else 
+			
 
 			if(!get_option('permalink_structure')){
 				return array("You have to set <strong><u><a href='".admin_url()."options-permalink.php"."'>permalinks</a></u></strong>", "error");
 			}else if($res = $this->checkSuperCache($path, $htaccess)){
 				return $res;
+			}else if($this->isPluginActive('wp-hide-security-enhancer/wp-hide.php')){
+				return array("WP Hide & Security Enhancer needs to be deactived<br>", "error");
 			}else if($this->isPluginActive('adrotate/adrotate.php') || $this->isPluginActive('adrotate-pro/adrotate.php')){
 				return $this->warningIncompatible("AdRotate");
 			}else if($this->isPluginActive('mobilepress/mobilepress.php')){
@@ -390,7 +393,7 @@
 
 
 			$data = "# BEGIN LBCWpFastestCache"."\n".
-					'<FilesMatch "\.(?i:ico|pdf|flv|jpg|jpeg|png|gif|js|css|swf|x-html|css|xml|js|woff|woff2|ttf|svg|eot)(\.gz)?$">'."\n".
+					'<FilesMatch "\.(ico|pdf|flv|jpg|jpeg|png|gif|js|css|swf|x-html|css|xml|js|woff|woff2|ttf|svg|eot)(\.gz)?$">'."\n".
 					'<IfModule mod_expires.c>'."\n".
 					'ExpiresActive On'."\n".
 					'ExpiresDefault A0'."\n".
@@ -701,6 +704,13 @@
 			$wpFastestCacheRenderBlockingCss = isset($this->options->wpFastestCacheRenderBlockingCss) ? 'checked="checked"' : "";
 
 			$wpFastestCacheLanguage = isset($this->options->wpFastestCacheLanguage) ? $this->options->wpFastestCacheLanguage : "eng";
+			
+
+			$wpFastestCacheLazyLoad = isset($this->options->wpFastestCacheLazyLoad) ? 'checked="checked"' : "";
+
+
+
+
 			$wpFastestCacheLBC = isset($this->options->wpFastestCacheLBC) ? 'checked="checked"' : "";
 			$wpFastestCacheLoggedInUser = isset($this->options->wpFastestCacheLoggedInUser) ? 'checked="checked"' : "";
 			$wpFastestCacheMinifyCss = isset($this->options->wpFastestCacheMinifyCss) ? 'checked="checked"' : "";
@@ -770,7 +780,7 @@
 							}else if((isset($_POST["wpFastestCachePage"])) && ("wpfc-".$_POST["wpFastestCachePage"] == $value["id"])){
 								$checked = ' checked="checked" ';
 							}
-							echo '<input '.$checked.' type="radio" id="'.$value["id"].'" name="tabGroup1">'."\n";
+							echo '<input '.$checked.' type="radio" id="'.$value["id"].'" name="tabGroup1" style="display:none;">'."\n";
 							echo '<label for="'.$value["id"].'">'.$value["title"].'</label>'."\n";
 						}
 					?>
@@ -1041,18 +1051,38 @@
 								<?php } ?>
 							<?php } ?>
 
-							<?php if(class_exists("WpFastestCachePowerfulHtml") && method_exists("WpFastestCachePowerfulHtml", "render_blocking")){ ?>
-							<div class="questionCon">
-								<div class="question">Render Blocking Js</div>
-								<div class="inputCon"><input type="checkbox" <?php echo $wpFastestCacheRenderBlocking; ?> id="wpFastestCacheRenderBlocking" name="wpFastestCacheRenderBlocking"><label for="wpFastestCacheRenderBlocking">Remove render-blocking JavaScript</label></div>
-								<div class="get-info"><a target="_blank" href="http://www.wpfastestcache.com/premium/render-blocking-js/"><img src="<?php echo plugins_url("wp-fastest-cache/images/info.png"); ?>" /></a></div>
-							</div>
+							<?php if(class_exists("WpFastestCachePowerfulHtml")){ ?> 
+								<?php if(method_exists("WpFastestCachePowerfulHtml", "render_blocking")){ ?>
+									<div class="questionCon">
+										<div class="question">Render Blocking Js</div>
+										<div class="inputCon"><input type="checkbox" <?php echo $wpFastestCacheRenderBlocking; ?> id="wpFastestCacheRenderBlocking" name="wpFastestCacheRenderBlocking"><label for="wpFastestCacheRenderBlocking">Remove render-blocking JavaScript</label> <b style="color:red;">(Beta)</b></div>
+										<div class="get-info"><a target="_blank" href="http://www.wpfastestcache.com/premium/render-blocking-js/"><img src="<?php echo plugins_url("wp-fastest-cache/images/info.png"); ?>" /></a></div>
+									</div>
+								<?php }else{ ?>
+									<div class="questionCon update-needed">
+										<div class="question">Render Blocking Js</div>
+										<div class="inputCon"><input type="checkbox" id="wpFastestCacheRenderBlocking" name="wpFastestCacheRenderBlocking"><label for="wpFastestCacheRenderBlocking">Remove render-blocking JavaScript</label> <b style="color:red;">(Beta)</b></div>
+										<div class="get-info"><a target="_blank" href="http://www.wpfastestcache.com/premium/render-blocking-js/"><img src="<?php echo plugins_url("wp-fastest-cache/images/info.png"); ?>" /></a></div>
+									</div>
+								<?php } ?>
 							<?php }else{ ?>
-							<div class="questionCon disabled">
-								<div class="question">Render Blocking Js</div>
-								<div class="inputCon"><input type="checkbox" <?php echo $wpFastestCacheRenderBlocking; ?> id="wpFastestCacheRenderBlocking" name="wpFastestCacheRenderBlocking"><label for="wpFastestCacheRenderBlocking">Remove render-blocking JavaScript</label></div>
+								<div class="questionCon disabled">
+									<div class="question">Render Blocking Js</div>
+									<div class="inputCon"><input type="checkbox" id="wpFastestCacheRenderBlocking" name="wpFastestCacheRenderBlocking"><label for="wpFastestCacheRenderBlocking">Remove render-blocking JavaScript</label> <b style="color:red;">(Beta)</b></div>
+									<div class="get-info"><a target="_blank" href="http://www.wpfastestcache.com/premium/render-blocking-js/"><img src="<?php echo plugins_url("wp-fastest-cache/images/info.png"); ?>" /></a></div>
+								</div>
+							<?php } ?>
+
+
+							<?php if(false){ ?>
+							<div class="questionCon">
+								<div class="question">Lazy Load</div>
+								<div class="inputCon"><input type="checkbox" <?php echo $wpFastestCacheLazyLoad; ?> id="wpFastestCacheLazyLoad" name="wpFastestCacheLazyLoad"><label for="wpFastestCacheLazyLoad">Lazy Load</label></div>
 								<div class="get-info"><a target="_blank" href="http://www.wpfastestcache.com/premium/render-blocking-js/"><img src="<?php echo plugins_url("wp-fastest-cache/images/info.png"); ?>" /></a></div>
 							</div>
+
+							<?php include(WPFC_MAIN_PATH."templates/lazy_load.php"); ?>
+
 							<?php } ?>
 
 							<div class="questionCon">
@@ -1245,6 +1275,11 @@
 								    					$tmp_array = array("schedule" => $event["schedule"],
 								    									   "prefix" => $tmp_std->prefix,
 								    									   "content" => $tmp_std->content);
+
+								    					if(isset($tmp_std->hour) && isset($tmp_std->minute)){
+								    						$tmp_array["hour"] = $tmp_std->hour;
+								    						$tmp_array["minute"] = $tmp_std->minute;
+								    					}
 							    					}else{
 							    						// old cronjob which is (wp_fastest_cache)
 							    						$tmp_array = array("schedule" => $event["schedule"],
@@ -1431,6 +1466,8 @@
 							    				</button>
 						    				<?php }else{ ?>
 							    				<form action="http://api.wpfastestcache.net/paypal/buypremium/" method="post">
+							    					<input type="hidden" name="wpfclang" value="<?php echo $this->options->wpFastestCacheLanguage; ?>">
+							    					<input type="hidden" name="bloglang" value="<?php echo get_bloginfo('language'); ?>">
 							    					<input type="hidden" name="hostname" value="<?php echo str_replace(array("http://", "www."), "", $_SERVER["HTTP_HOST"]); ?>">
 								    				<button id="wpfc-buy-premium-button" type="submit" class="wpfc-btn primaryCta" style="width:200px;">
 								    					<span>Buy</span>
@@ -1781,6 +1818,11 @@
 			</div>
 
 			<div class="omni_admin_sidebar">
+				<div style="padding:0 !important;float:left;">
+					<a href="//partners.hostgator.com/c/149801/215624/3094" target="_blank">
+						<img src="<?php echo plugins_url("wp-fastest-cache/images/ads/hostgator-250x250.gif"); ?>" border="0" alt="" width="222" height="220"/>
+					</a>
+				</div>
 				<div class="omni_admin_sidebar_section" id="vote-us">
 					<h3 style="color: antiquewhite;">Rate Us</h3>
 					<ul>
@@ -1921,8 +1963,22 @@
 				</script>
 			<?php }else{ ?>
 				<script type="text/javascript">
-					jQuery(".update-needed").click(function(e){
-						Wpfc_Dialog.dialog("wpfc-modal-updatenow");
+					jQuery(".update-needed").click(function(){
+						if(jQuery("div[id^='wpfc-modal-updatenow-']").length === 0){
+							Wpfc_New_Dialog.dialog("wpfc-modal-updatenow", {close: function(){
+								Wpfc_New_Dialog.clone.find("div.window-content input").each(function(){
+									if(jQuery(this).attr("checked")){
+										var id = jQuery(this).attr("action-id");
+										jQuery("div.tab1 div[template-id='wpfc-modal-updatenow'] div.window-content input#" + id).attr("checked", true);
+									}
+								});
+
+								Wpfc_New_Dialog.clone.remove();
+							}});
+
+							Wpfc_New_Dialog.show_button("close");
+						}
+
 						return false;
 					});
 				</script>