+++ plugins/wpml-string-translation/classes/upgrade/interface-iwpml_st_upgrade_command.php
@@ -7,4 +7,6 @@ interface IWPML_St_Upgrade_Command {
 	public function run_ajax();
 
 	public function run_frontend();
+
+	public static function get_commnand_id();
 }