+++ plugins/wpml-string-translation/plugin.php
@@ -5,7 +5,7 @@ Plugin URI: https://wpml.org/
 Description: Adds theme and plugins localization capabilities to WPML | <a href="https://wpml.org">Documentation</a> | <a href="https://wpml.org/version/wpml-3-5-0/">WPML 3.5.0 release notes</a>
 Author: OnTheGoSystems
 Author URI: http://www.onthegosystems.com/
-Version: 2.4.1
+Version: 2.4.1.1
 Plugin Slug: wpml-string-translation
 */
 
@@ -13,7 +13,7 @@ if ( defined( 'WPML_ST_VERSION' ) ) {
 	return;
 }
 
-define( 'WPML_ST_VERSION', '2.4.1' );
+define( 'WPML_ST_VERSION', '2.4.1.1' );
 
 // Do not uncomment the following line!
 // If you need to use this constant, use it in the wp-config.php file