+++ plugins/sitepress-multilingual-cms/changelog.md
@@ -1,3 +1,8 @@
+# 3.5.1.1
+
+# Fixes
+* [wpmlcore-3303] Removed updating of ST DB Cache tables: moved to ST plugin, to avoid dependency issues
+
 # 3.5.1
 
 # Fixes